fixed url to svg images in CSS assets
Previously this was using the absolute URI `/themes/mandelbrot` to reference
svg images, but this only worked correctly when the fractal instance was
running under the root path. For deployed fractal sites, the pattern library
instance might be deployed under a subdirectory, in which case the Icon SVGs
would not be correctly found.

This also updates various npm dependencies and generates a new package-lock.json
to update transitive dependencies.